requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-WTF==1.2.1
WTForms==3.1.1
Werkzeug==3.0.1
email-validator==2.1.0
PyMySQL==1.1.0
mysql-connector-python==8.2.0
